pyproject.toml
[project]
name = "brokeout-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorama>=0.4.6",
    "discord-rpc>=5.1",
    "moderngl>=5.12.0",
    "munch>=4.0.0",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "pyaml>=25.7.0",
    "pygame-ce>=2.5.5",
]
